open-digital-platform-2_0/5G_Emulator_API/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-jaeger
opentelemetry-exporter-prometheus
opentelemetry-exporter-otlp
fastapi
requests
psutil
prometheus_client
uvicorn